# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AssignmentIn": {
        "properties": {
          "agentRole": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agentrole"
          },
          "phase": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phase"
          },
          "profileId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profileid"
          },
          "projectId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Projectid"
          }
        },
        "required": [
          "profileId"
        ],
        "title": "AssignmentIn",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PolicyIn": {
        "properties": {
          "allowedUpstreamProviders": {
            "anyOf": [
              {
                "items": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Allowedupstreamproviders"
          },
          "deniedUpstreamProviders": {
            "items": {
              "maxLength": 100,
              "minLength": 1,
              "type": "string"
            },
            "title": "Deniedupstreamproviders",
            "type": "array"
          },
          "denyDataCollection": {
            "default": true,
            "title": "Denydatacollection",
            "type": "boolean"
          },
          "openrouterAllowed": {
            "default": true,
            "title": "Openrouterallowed",
            "type": "boolean"
          },
          "requireZdr": {
            "default": false,
            "title": "Requirezdr",
            "type": "boolean"
          }
        },
        "title": "PolicyIn",
        "type": "object"
      },
      "PolicyOut": {
        "properties": {
          "allowedUpstreamProviders": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Allowedupstreamproviders"
          },
          "deniedUpstreamProviders": {
            "items": {
              "type": "string"
            },
            "title": "Deniedupstreamproviders",
            "type": "array"
          },
          "denyDataCollection": {
            "title": "Denydatacollection",
            "type": "boolean"
          },
          "openrouterAllowed": {
            "title": "Openrouterallowed",
            "type": "boolean"
          },
          "requireZdr": {
            "title": "Requirezdr",
            "type": "boolean"
          }
        },
        "required": [
          "openrouterAllowed",
          "allowedUpstreamProviders",
          "deniedUpstreamProviders",
          "requireZdr",
          "denyDataCollection"
        ],
        "title": "PolicyOut",
        "type": "object"
      },
...
      "ResolvedOut": {
        "properties": {
          "profileId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profileid"
          }
        },
        "required": [
          "profileId"
        ],
        "title": "ResolvedOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "nexti_api__ai__assignments__AssignmentOut": {
        "properties": {
          "agentRole": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agentrole"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "phase": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phase"
          },
          "profileId": {
            "format": "uuid",
            "title": "Profileid",
            "type": "string"
          },
          "projectId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Projectid"
          }
        },
        "required": [
          "id",
          "projectId",
          "phase",
          "agentRole",
          "profileId"
        ],
        "title": "AssignmentOut",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "NexTI Platform API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/ai/assignments": {
      "get": {
        "description": "Tenant-wide rows, plus the rows of `projectId` when given.",
        "operationId": "list_assignments_api_v1_ai_assignments_get",
        "parameters": [
          {
            "in": "query",
            "name": "projectId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Projectid"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/nexti_api__ai__assignments__AssignmentOut"
                  },
                  "title": "Response List Assignments Api V1 Ai Assignments Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Assignments",
        "tags": [
          "ai"
        ]
      },
      "put": {
        "operationId": "set_assignment_api_v1_ai_assignments_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignmentIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/nexti_api__ai__assignments__AssignmentOut"
                  },
                  "title": "Response Set Assignment Api V1 Ai Assignments Put",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Assignment",
        "tags": [
          "ai"
        ]
      }
    },
    "/api/v1/ai/assignments:resolve": {
      "get": {
        "description": "Which profile the gateway would use for this context (the same cascade as the gateway).",
        "operationId": "resolve_api_v1_ai_assignments_resolve_get",
        "parameters": [
          {
            "in": "query",
            "name": "projectId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Projectid"
            }
          },
          {
            "in": "query",
            "name": "phase",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Phase"
            }
          },
          {
            "in": "query",
            "name": "agentRole",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Agentrole"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolvedOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolve",
        "tags": [
          "ai"
        ]
      }
    },
...
    "/api/v1/ai/policy": {
      "get": {
        "operationId": "get_policy_api_v1_ai_policy_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Policy",
        "tags": [
          "ai"
        ]
      },
      "put": {
        "operationId": "set_policy_api_v1_ai_policy_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PolicyIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Policy",
        "tags": [
          "ai"
        ]
      }
    },
...
  }
}